#25 文字色の対応
